chore(eth-lc): fix signature verification
