+++ cosmwasm/union-ibc/light-clients/ethereum/Cargo.toml
@@ -19,6 +19,7 @@ crate-type = ["cdylib", "rlib"]
 alloy                       = { workspace = true, features = ["sol-types"] }
 beacon-api-types            = { workspace = true }
 bytes                       = { workspace = true }
+cometbls-light-client-types = { workspace = true, features = ["proto"] }
 cosmwasm-schema             = { version = "2.1.4" }
 cosmwasm-std                = { version = "2.1.4", features = ["abort"] }
 cw-storage-plus             = { version = "2.0.0" }
@@ -39,10 +40,9 @@ sha3                        = { workspace = true }
 thiserror                   = { workspace = true }
 tiny-keccak                 = { workspace = true, features = ["keccak"] }
 union-ibc                   = { workspace = true, features = ["library"] }
-union-ibc-light-client      = { workspace = true, default-features = false, features = [ "cosmwasm-2" ] }
+union-ibc-light-client      = { workspace = true, default-features = false, features = ["cosmwasm-2"] }
 union-ibc-msg               = { workspace = true }
 unionlabs                   = { workspace = true, features = ["ethabi"] }
-cometbls-light-client-types = { workspace = true, features = ["proto"] }
 
 [dev-dependencies]
 base64      = { workspace = true }
