+++ cosmwasm/union-ibc/light-clients/ethereum/src/client.rs
@@ -15,15 +15,11 @@ use evm_storage_verifier::{
 use union_ibc_light_client::IbcClientCtx;
 use union_ibc_msg::lightclient::Status;
 use unionlabs::{
-    encoding::{DecodeAs, Proto},
-    ensure,
-    ethereum::ibc_commitment_key,
-    hash::H256,
-    ibc::core::client::height::Height,
-    uint::U256,
+    encoding::Proto, ensure, ethereum::ibc_commitment_key, hash::H256,
+    ibc::core::client::height::Height, uint::U256,
 };
 
-use crate::{custom_query::VerificationContext, errors::Error};
+use crate::{errors::Error, verification::VerificationContext};
 
 pub enum EthereumLightClient {}
 
@@ -1045,10 +1041,3 @@ pub fn verify_misbehaviour<C: ChainSpec>(
 //         }
 //     }
 // }
-
-#[test]
-fn test_update() {
-    let client_message = hex_literal::hex!("000000000000000000000000000000000000000000000000000000000000121e00000000000000000000000000000000000000000000000000000000673e5f09000000000000000000000000000000000000000000000000000000001678f3752f4975ab7e75a677f43efebf53e0ec05460d2cf55506ad08d6b05254f96a500d2f4975ab7e75a677f43efebf53e0ec05460d2cf55506ad08d6b05254f96a500d121916d2ccd9d1e831d4bba7333b22130cc71592ce5976b91cf1ee7a212a5a87000000000000000000000000000000000000000000000000000000000000103d0000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000018002d7784e0777d028503c72a1d7f854c4fe5e87fcd4e6792f0b3eefbb1d64ec95149a1c4be3204cf019d678c656b4ddab73185a704852dabef605be0bfad006a8071f587e2dd229db96ba3d022df762b42c166dbdcd0d363f35d411aeec3068d8009f1f4506591fd95f3f8d21b1e29f1c48738bf670f55ddd450d5070540e103d05e4bc2ef4d090ef06fcae2873e242e3bc02d26a5d2625b4a38d765505e7b5cb1a4cc5476bb9b43f4a2812d53a81183af71e9c437627fe91281adeddc7db19932117d5816f60344878430d900070abf2102ec8a9cb73b5c66c5933a79e0a1ef101bd374756d2bfa07f9e1adcc136236a13b261c5dcfd86977421a3dffcc5550f0a244429e18a6162ff9299b09bfcc878f4d386c1c62e8788103b43b0fdfbd3ab03fc8dd2b4e69444c27f3a3120aa7875c10cb67b1d65f079d4b1be92ca91ea1a23ee9387988930d9160d962d779006ea0f8a93d869149ff809fe557181c7f6462f119c0672321a4a5b5c47ffe8faeae4d209d2a49e0ef392ac307c4e6880a4b8");
-    let client_message =
-        cometbls_light_client_types::header::Header::decode_as::<Proto>(&client_message).unwrap();
-}
