+++ cosmwasm/union-ibc/light-clients/ethereum/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod client;
 #[cfg(any(test, not(feature = "library")))]
 pub mod contract;
-pub mod custom_query;
 pub mod errors;
+pub mod verification;
